# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CNSTRCT API MCP Server",
    "description": "CNSTRCT API for integrating with QuickBooks Online for transactions, expenses, and invoices",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/qbo/callback": {
      "get": {
        "summary": "OAuth callback",
        "description": "Callback endpoint for OAuth2 flow",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "State parameter",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "realmId",
            "in": "query",
            "description": "QuickBooks Company ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth token response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/qbo/retrieveToken": {
      "get": {
        "summary": "Retrieve current OAuth token",
        "description": "Returns the current OAuth token",
        "responses": {
          "200": {
            "description": "OAuth token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No token found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/qbo/refreshAccessToken": {
      "get": {
        "summary": "Refresh OAuth access token",
        "description": "Refreshes the OAuth access token",
        "responses": {
          "200": {
            "description": "Refreshed OAuth token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company": {
      "get": {
        "summary": "Get company information",
        "description": "Returns information about the QuickBooks company",
        "responses": {
          "200": {
            "description": "Company information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts": {
      "get": {
        "summary": "Get all accounts",
        "description": "Returns a list of QuickBooks accounts",
        "parameters": [
          {
            "name": "AccountType",
            "in": "query",
            "description": "Filter by account type",
            "schema": {
              "type": "string",
              "enum": [
                "Bank",
                "Other Current Asset",
                "Fixed Asset",
                "Other Asset",
                "Accounts Receivable",
                "Equity",
                "Expense",
                "Other Expense",
                "Income",
                "Other Income",
                "Cost of Goods Sold",
                "Accounts Payable",
                "Credit Card",
                "Long Term Liability",
                "Other Current Liability"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of records to return",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "fetchAll",
            "in": "query",
            "description": "Fetch all records",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new account",
        "description": "Creates a new QuickBooks account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "Name",
                  "AccountType"
                ],
                "properties": {
                  "Name": {
                    "type": "string",
                    "description": "Account name"
                  },
                  "AccountType": {
                    "type": "string",
                    "description": "Type of account",
                    "enum": [
                      "Bank",
                      "Other Current Asset",
                      "Fixed Asset",
                      "Other Asset",
                      "Accounts Receivable",
                      "Equity",
                      "Expense",
                      "Other Expense",
                      "Income",
                      "Other Income",
                      "Cost of Goods Sold",
                      "Accounts Payable",
                      "Credit Card",
                      "Long Term Liability",
                      "Other Current Liability"
                    ]
                  },
                  "AccountSubType": {
                    "type": "string",
                    "description": "Account sub-type"
                  },
                  "Description": {
                    "type": "string",
                    "description": "Account description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "summary": "Get a specific account",
        "description": "Returns a specific QuickBooks account by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an account",
        "description": "Updates a specific QuickBooks account by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Account ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string",
                    "description": "Account name"
                  },
                  "Description": {
                    "type": "string",
                    "description": "Account description"
                  },
                  "Active": {
                    "type": "boolean",
                    "description": "Whether the account is active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated account",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/reports/profit-loss": {
      "get": {
        "summary": "Profit and Loss report",
        "description": "Retrieves a Profit and Loss report for the specified date range and departments",
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date for the report"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date for the report"
          },
          {
            "in": "query",
            "name": "departments",
            "schema": {
              "type": "string"
            },
            "description": "Department(s) to filter the report by"
          }
        ],
        "responses": {
          "200": {
            "description": "Profit and Loss report",
            "content": {
              "application/json": {
                
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                
              }
            }
          }
        }
      }
    },
...
  }
}